requirements.txt
torch>=2.0
torchvision
regex
ftfy
tqdm
huggingface_hub
safetensors
timm
onnx
onnxruntime-gpu
pandas
tabulate
